refactor: Product component
